ai-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
python-multipart>=0.0.6
pydantic>=2.5.0
nltk>=3.8.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openai>=1.6.0
python-dotenv>=1.0.0
joblib>=1.3.0
groq>=0.4.0
requests